11-Final-Project/docker/etl/requirements.txt
pymongo
dnspython
sqlalchemy
psycopg2-binary
vaderSentiment